+++ examples/right_spinner.py
@@ -13,15 +13,15 @@ from yaspin import yaspin
 
 
 def main():
-    with yaspin(text="Right spinner", right=True) as sp:
+    with yaspin(text="Right spinner", side="right", color="cyan") as sp:
         time.sleep(2)
 
         # Switch to left spinner
-        sp.right = False
+        sp.side = "left"  # or just ``sp.left``
         sp.text = "Left spinner"
 
         time.sleep(2)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()