+++ tests/conftest.py
@@ -7,30 +7,33 @@ tests.conftest
 Tests data.
 """
 
+import signal
+import sys
+
 import pytest
 
-from yaspin.termcolor import colored
+from yaspin.compat import iteritems
+from yaspin.constants import COLOR_MAP
+from yaspin.signal_handlers import default_handler, fancy_handler
 
 
 frame_cases = [
     # XXX: try byte strings
     # String types
-    "",                                            # empty
-    "+x*",                                         # ascii str
-    "⢄⢂⢁⡁⡈⡐⡠",                                     # non-ascii in str
-    u"⢹⢺⢼⣸",                                       # non-ascii in unicode str
-
+    "",  # empty
+    "+x*",  # ascii str
+    "⢄⢂⢁⡁⡈⡐⡠",  # non-ascii in str
+    u"⢹⢺⢼⣸",  # non-ascii in unicode str
     # Lists
-    [],                                            # List[]
-    [b"\xf0\x9f\x8c\xb2", b"\xf0\x9f\x8e\x84"],    # List[bytes]
-    [u"🌲", u"🎄"],                                # List[unicode]
-    ["⢹", "⢺", "⢼", "⣸"],                          # List[str], non-ascii
-
+    [],  # List[]
+    [b"\xf0\x9f\x8c\xb2", b"\xf0\x9f\x8e\x84"],  # List[bytes]
+    [u"🌲", u"🎄"],  # List[unicode]
+    ["⢹", "⢺", "⢼", "⣸"],  # List[str], non-ascii
     # Tuples
-    (),                                            # Tuple[]
-    (b"\xf0\x9f\x8c\xb2", b"\xf0\x9f\x8e\x84"),    # Tuple[bytes]
-    (u"🌲", u"🎄"),                                # Tuple[unicode]
-    ("⢹", "⢺", "⢼", "⣸"),                          # Tuple[str], non-ascii
+    (),  # Tuple[]
+    (b"\xf0\x9f\x8c\xb2", b"\xf0\x9f\x8e\x84"),  # Tuple[bytes]
+    (u"🌲", u"🎄"),  # Tuple[unicode]
+    ("⢹", "⢺", "⢼", "⣸"),  # Tuple[str], non-ascii
 ]
 
 
@@ -40,13 +43,11 @@ frame_ids = [
     "'ascii str'",
     "'non-ascii str'",
     "'non-ascii unicode str'",
-
     # Lists
     "'List[]'",
     "'List[bytes]'",
     "'List[unicode]'",
     "'List[str] non-ascii'",
-
     # Tuples
     "'Tuple[]'",
     "'Tuple[bytes]'",
@@ -57,11 +58,10 @@ frame_ids = [
 
 text_cases = [
     # XXX: try byte strings
-
-    "",             # empty
-    "Loading",      # ascii str
-    "ℙƴ☂ℌøἤ",       # non-ascii in str
-    u"Загрузка",    # non-ascii in unicode str
+    "",  # empty
+    "Loading",  # ascii str
+    "ℙƴ☂ℌøἤ",  # non-ascii in str
+    u"Загрузка",  # non-ascii in unicode str
 ]
 
 
@@ -88,21 +88,15 @@ def text(request):
     return request.param
 
 
-@pytest.fixture(
-    scope="session",
-    params=[False, True],
-    ids=["'left'", "'right'"]
-)
-def right(request):
+@pytest.fixture(scope="session", params=["left", "right"])
+def side(request):
     return request.param
 
 
 @pytest.fixture(
-    scope="session",
-    params=[False, True],
-    ids=["default", "reverse"]
+    scope="session", params=[False, True], ids=["default", "reversal"]
 )
-def reverse(request):
+def reversal(request):
     return request.param
 
 
@@ -120,72 +114,170 @@ def color_id_func(case):
     return val
 
 
-@pytest.fixture(scope="session", ids=color_id_func, params=[
-    # Empty values
-    ("", ""),
-    (None, None),
+def attrs_id_func(case):
+    if isinstance(case, list):
+        val = ", ".join(case)
+    if isinstance(case, tuple):
+        attrs, _ = case
+        val = ", ".join(attrs)
+    return val
 
-    # Supported text colors
-    ("red", "red"),
-    ("green", "green"),
-    ("yellow", "yellow"),
-    ("blue", "blue"),
-    ("magenta", "magenta"),
-    ("cyan", "cyan"),
-    ("white", "white"),
 
-    # Unsupported text colors
-    ("black", ValueError()),
-    ("brown", ValueError()),
-    ("orange", ValueError()),
-
-    # Uppercase handling
-    ("Red", "red"),
-    ("grEEn", "green"),
-    ("BlacK", ValueError()),
-
-    # Callables
-    (
-        lambda frame: colored(frame, 'red', attrs=['bold']),
-        lambda frame: colored(frame, 'red', attrs=['bold']),
-    ),
-])
-def colors_test_cases(request):
+@pytest.fixture(
+    scope="session",
+    ids=color_id_func,
+    params=[
+        # Empty values
+        ("", ""),
+        (None, None),
+        # Supported text colors
+        ("red", "red"),
+        ("green", "green"),
+        ("yellow", "yellow"),
+        ("blue", "blue"),
+        ("magenta", "magenta"),
+        ("cyan", "cyan"),
+        ("white", "white"),
+        # Unsupported text colors
+        ("Red", ValueError()),
+        ("orange", ValueError()),
+    ],
+)
+def color_test_cases(request):
     return request.param
 
 
-@pytest.fixture(scope="session", ids=color_id_func, params=[
-    "red",
-    "green",
-    "yellow",
-    "blue",
-    "magenta",
-    "cyan",
-    "white",
-    lambda frame: colored(frame, 'red', attrs=['bold']),
-])
+@pytest.fixture(
+    scope="session",
+    ids=color_id_func,
+    params=[
+        # Empty values
+        ("", ""),
+        (None, None),
+        # Supported highlights
+        ("on_red", "on_red"),
+        ("on_green", "on_green"),
+        ("on_yellow", "on_yellow"),
+        ("on_blue", "on_blue"),
+        ("on_magenta", "on_magenta"),
+        ("on_cyan", "on_cyan"),
+        ("on_white", "on_white"),
+        # Unsupported highlights
+        ("on_foo", ValueError()),
+    ],
+)
+def on_color_test_cases(request):
+    return request.param
+
+
+@pytest.fixture(
+    scope="session",
+    ids=attrs_id_func,
+    params=[
+        # Supported attrs
+        (["bold"], ["bold"]),
+        (["dark"], ["dark"]),
+        (["underline"], ["underline"]),
+        (["blink"], ["blink"]),
+        (["reverse"], ["reverse"]),
+        (["concealed"], ["concealed"]),
+        # Multiple attrs
+        (["bold", "dark"], ["bold", "dark"]),
+        (["bold", "dark", "reverse"], ["bold", "dark", "reverse"]),
+        # Unsupported attrs
+        (["Dark"], ValueError()),
+        (["bold", "bar"], ValueError()),
+    ],
+)
+def attrs_test_cases(request):
+    return request.param
+
+
+@pytest.fixture(
+    scope="session",
+    ids=color_id_func,
+    params=sorted([k for k, v in iteritems(COLOR_MAP) if v == "color"]),
+)
 def supported_colors(request):
     return request.param
 
 
-@pytest.fixture(scope="session", params=[
-    # Empty
-    b"",
-    u"",
+@pytest.fixture(
+    scope="session",
+    ids=color_id_func,
+    params=sorted([k for k, v in iteritems(COLOR_MAP) if v == "on_color"]),
+)
+def supported_highlights(request):
+    return request.param
 
-    # Success
-    b"OK",
-    u"OK",
-    b"\xe2\x9c\x94",
-    u"✔",
 
-    # Sun
-    b"\xe2\x98\x80\xef\xb8\x8f",
-    u"☀️",
+@pytest.fixture(
+    scope="session",
+    ids=attrs_id_func,
+    params=sorted(
+        [[k] for k, v in iteritems(COLOR_MAP) if v == "attrs"]
+        + [
+            ["bold", "dark"],
+            ["blink", "concealed", "reverse"],
+            ["underline", "concealed", "bold", "dark", "blink", "reverse"],
+        ]
+    ),
+)
+def supported_attrs(request):
+    return request.param
 
-    # Spark
-    b"\xf0\x9f\x92\xa5",
-    u"💥",
-])
+
+@pytest.fixture(
+    scope="session",
+    params=[
+        # Empty
+        b"",
+        u"",
+        # Success
+        b"OK",
+        u"OK",
+        b"\xe2\x9c\x94",
+        u"✔",
+        # Sun
+        b"\xe2\x98\x80\xef\xb8\x8f",
+        u"☀️",
+        # Spark
+        b"\xf0\x9f\x92\xa5",
+        u"💥",
+    ],
+)
 def final_text(request):
+    return request.param
+
+
+@pytest.fixture(
+    scope="session",
+    params=[
+        None,
+        {signal.SIGUSR1: signal.SIG_DFL},
+        {signal.SIGTERM: signal.SIG_IGN},
+        {signal.SIGTERM: signal.default_int_handler},
+        {signal.SIGHUP: default_handler},
+        {signal.SIGINT: fancy_handler},
+        {signal.SIGINT: lambda signum, frame: sys.exit(1)},
+        {
+            signal.SIGUSR1: signal.SIG_DFL,
+            signal.SIGTERM: signal.SIG_IGN,
+            signal.SIGHUP: default_handler,
+            signal.SIGINT: fancy_handler,
+            signal.SIGINT: lambda signum, frame: sys.exit(1),
+        },
+    ],
+    ids=[
+        "no sigmap",
+        "SIGUSR1 - SIG_DFL",
+        "SIGTERM - SIG_IGN",
+        "SIGTERM - default_int_handler",
+        "SIGHUP - default_handler",
+        "SIGINT - fancy_handler",
+        "SIGINT - custom handler",
+        "Multiple signals-handlers map",
+    ],
+)
+def sigmap_test_cases(request):
     return request.param